replace client with simpleClient
